Move get_nc_files to io.py
