+++ external/fv3fit/fv3fit/emulation/data/__init__.py
@@ -1,3 +1,4 @@
 from .config import TransformConfig
 from . import transforms
-from .load import get_nc_files, batched_to_tf_dataset
+from .load import batched_to_tf_dataset
+from .io import get_nc_files
