+++ external/fv3fit/fv3fit/emulation/data/load.py
@@ -1,26 +1,9 @@
 import logging
-import os
 import tensorflow as tf
-from typing import Callable, List, Sequence
-
-from vcm import get_fs
+from typing import Callable, Sequence
 
 
 logger = logging.getLogger(__name__)
-
-
-def get_nc_files(path: str) -> List[str]:
-    """
-    Get a list of netCDF files from a remote/local directory
-
-    Args:
-        path: Local or remote gcs path to netCDF directory
-    """
-
-    fs = get_fs(path)
-    files = list(fs.glob(os.path.join(path, "*.nc")))
-
-    return files
 
 
 def batched_to_tf_dataset(
